refactor: rename directory to follow python structure
