+++ fullstack/js/price-calculator/app/__tests__/calculate.spec.js
