+++ fullstack/js/price-calculator/app/calculate.js
