+++ fullstack/js/price-calculator/app/utils/__tests__/combinations.js
