+++ fullstack/js/price-calculator/app/utils/__tests__/map-prices.js
