+++ fullstack/js/price-calculator/app/utils/combinations.js
