+++ fullstack/js/price-calculator/app/utils/map-prices.js
