+++ fullstack/js/price-calculator/bin/price-calculator
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
 require = require('esm')(module /*, options*/)
-require('../cli').calculatePriceForCart(process.argv)
+require('../main').calculatePriceForCart(process.argv)
